Forked go-shopify, in order to convert internal ids to int64 from straight int.  Builds, needs testing to verify before pushing a pull request back to the bold-commerce folks

+++ inventory_item_test.go
@@ -12,7 +12,7 @@ func inventoryItemTests(t *testing.T, item *InventoryItem) {
 		return
 	}
 
-	expectedInt := 808950810
+	expectedInt := int64(808950810)
 	if item.ID != expectedInt {
 		t.Errorf("InventoryItem.ID returned %+v, expected %+v", item.ID, expectedInt)
 	}
